+++ CHANGELOG.md
@@ -7,3 +7,4 @@ Changelog for 6.0
 *   Bumped all dependencies. PHP 7.2+ and Symfony 4.2+ are now required.
 *   Added `classnames()` twig function.
 *   Added a custom `Profiler` wrapper, that is always available and allows to disable symfony's profiler.
+*   Added a `data_container()` twig function.